% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{psMarker_Top}
\alias{psMarker_Top}
\title{Select top pseudo-bulk marker genes from MiCBuS DE results}
\usage{
psMarker_Top(DEpseudo.df, ntop = 200)
}
\arguments{
\item{DEpseudo.df}{Data frame returned by \code{MiCBuS()}.}

\item{ntop}{Integer. Number of top marker genes to return.}
}
\value{
Character vector of top-ranked marker genes.
}
\description{
Ranks and returns the top markers.
}
