% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiCBuS.R
\name{simulate_bulk_dir}
\alias{simulate_bulk_dir}
\title{Generate Dirichlet-based pseudo-bulk RNA-seq samples}
\usage{
simulate_bulk_dir(nsamples, prop, s = 10, ct, ds, ncells = 1000)
}
\arguments{
\item{nsamples}{Number of Dirichlet-based pseudo-bulk RNA-seq samples to generate}

\item{prop}{Initial proportions of cell types}

\item{s}{A parameter to adjust the spread around initial proportions}

\item{ct}{Cell types existed in the scRNA-seq data}

\item{ds}{An object prepared for SimBu::simulate_bulk()}

\item{ncells}{Number of cells for each sample}
}
\value{
A list
}
\description{
Generate Dirichlet-based pseudo-bulk RNA-seq samples
}
